Better return handling

+++ web/api/delete_whitelist_entry.py
@@ -5,8 +5,22 @@ import sys
 import boto3
 
 
+def get_return(code, body):
+    return {
+        "statusCode": code,
+        "headers": {
+            "Access-Control-Allow-Origin": "*",
+            "Access-Control-Allow-Credentials": True,
+        },
+        "body": json.dumps(body),
+    }
+
+
 def lambda_handler(event, context):
     client = boto3.client("dynamodb")
+
+    if event.get("resource_id") in (None, ""):
+        return get_return(400, "Resource ID cannot be null.")
 
     try:
         response = client.delete_item(
@@ -16,19 +30,6 @@ def lambda_handler(event, context):
             },
         )
 
-        return {
-            "statusCode": response["ResponseMetadata"]["HTTPStatusCode"],
-            "headers": {
-                "Access-Control-Allow-Origin": "*",
-                "Access-Control-Allow-Credentials": True,
-            },
-        }
+        return get_return(response["ResponseMetadata"]["HTTPStatusCode"], None)
     except:
-        return {
-            "statusCode": 500,
-            "headers": {
-                "Access-Control-Allow-Origin": "*",
-                "Access-Control-Allow-Credentials": True,
-            },
-            "body": sys.exc_info()[1],
-        }
+        return get_return(400, sys.exc_info()[1])
